[8.x] feat(phpstan): support levels 3, 4 and 5 (#2095)
* wip

* feat(phpstan): support level 3

* feat(phpstan): support level 4

* feat(phpstan): support level 5

* fix(phpstan): fix Money constructor

* some fixes after merge

* fix time during tests

+++ app/Services/FareService.php
@@ -50,7 +50,7 @@ class FareService extends Service
             /**
              * See if there's match with the provided fares, so we can copy the information over
              *
-             * @var PirepFare $pirep_fare
+             * @var ?PirepFare $pirep_fare
              */
             $pirep_fare = $fares->where('fare_id', $fare->id)->first();
 
@@ -232,9 +232,9 @@ class FareService extends Service
 
         if (filled($pivot->capacity)) {
             if (strpos($pivot->capacity, '%', -1) !== false) {
-                $fare->capacity = floor(Math::getPercent($fare->capacity, $pivot->capacity));
+                $fare->capacity = (int) floor(Math::getPercent($fare->capacity, $pivot->capacity));
             } else {
-                $fare->capacity = floor($pivot->capacity);
+                $fare->capacity = (int) floor($pivot->capacity);
             }
         }
